firmware: qcom_scm: Permanently enable clocks at probe time
For some reason the clocks were only ever enabled in some (but not
all!) scm calls.. This led to a bogus enable_count (though the clocks
themselves were on either from the SMD_/RPM clock drivers or the
bootloader), which prevents the addition of rpm clock suspend.

Turn the clocks on at probe and leave them on the entire time. *Actually*
putting them to sleep would cause unimaginable chaos, anyway.

Signed-off-by: Konrad Dybcio <konrad.dybcio@linaro.org>